Pass ICC building.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@6761 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/DuetPkg/PciRootBridgeNoEnumerationDxe/Ia32/PcatIo.c
@@ -555,7 +555,7 @@ SaveCommandRegister (
   //
   // Clear the memory enable bit
   //
-  Command = Context->CommandRegisterBuffer[Index] & (~0x02);
+  Command = (UINT16) (Context->CommandRegisterBuffer[Index] & (~0x02));
 
   IoDev->Pci.Write (IoDev, EfiPciWidthUint16, Address, 1, &Command);
 }
